Added-payment-discount-receipt

+++ POSWinforms/Models/OrderDetail.cs
@@ -14,9 +14,9 @@ namespace POSWinforms.Models
         public string ItemDescription { get; set; }
         public int Quantity { get; set; }
 
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public int Discount { get; set; }
-        public double Total { get; set; }
+        public decimal Total { get; set; }
         public long OrderID { get; set; }
     }
 }
